+++ assets/javascripts/kitten/components/cards/backing-card/components/amount.js
@@ -1,0 +1,20 @@
+import React from 'react'
+import classNames from 'classnames'
+import { Text } from '../../../../index'
+
+export const Amount = props => {
+  return (
+    <Text
+      tag="p"
+      weight="bold"
+      lineHeight="1"
+      fontStyle="italic"
+      {...props}
+      className={classNames(
+        'k-BackingCard__amount',
+        'k-BackingCard__drawer',
+        props.className,
+      )}
+    />
+  )
+}