+++ assets/javascripts/kitten/components/cards/backing-card/components/image.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import classNames from 'classnames'
+
+export const Image = props => {
+  return (
+    <div
+      {...props}
+      className={classNames('k-BackingCard__imageWrapper', props.className)}
+    />
+  )
+}