+++ assets/javascripts/kitten/components/cards/backing-card/components/index.js
@@ -1,0 +1,7 @@
+export { Info } from './info'
+export { Title } from './title'
+export { Image } from './image'
+export { Amount } from './amount'
+export { Button } from './button'
+export { TagList } from './tag-list'
+export { Description } from './description'