+++ assets/javascripts/kitten/components/cards/backing-card/index.js
@@ -1,0 +1,48 @@
+import React from 'react'
+import classNames from 'classnames'
+import { StyledBackingCard } from './styles'
+import {
+  Info,
+  Title,
+  Image,
+  Amount,
+  Button,
+  TagList,
+  Description,
+} from './components'
+
+export const BackingCard = ({ children, className, disabled, ...props }) => {
+  return (
+    <StyledBackingCard
+      {...props}
+      className={classNames('k-BackingCard', className, {
+        'k-BackingCard--disabled': disabled,
+      })}
+    >
+      {React.Children.map(children, child => {
+        if (!child) return null
+        return child.type.name === 'Image' ? child : null
+      })}
+
+      <div className="k-BackingCard__gridWrapper">
+        {React.Children.map(children, child => {
+          if (!child) return null
+          return ['Image', 'Button'].includes(child.type.name) ? null : child
+        })}
+      </div>
+
+      {React.Children.map(children, child => {
+        if (!child) return null
+        return child.type.name === 'Button' ? child : null
+      })}
+    </StyledBackingCard>
+  )
+}
+
+BackingCard.Info = Info
+BackingCard.Title = Title
+BackingCard.Image = Image
+BackingCard.Amount = Amount
+BackingCard.Button = Button
+BackingCard.TagList = TagList
+BackingCard.Description = Description